% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oatk.R
\name{bc}
\alias{bc}
\title{The Knockoff Filter}
\usage{
bc(y, X, q = 0.1, offset = 0)
}
\arguments{
\item{y}{Response variable}

\item{X}{Design matrix}

\item{q}{The desired FDR level}

\item{offset}{Offset parameter when estimated the FDR. Setting c=1 results in a more conservative FDR result.}
}
\value{
A list containing the vector rej, which are the indices of the columns of X corresponding to the non-null variables,
the vector W which are the test statistics for each variable, and the numeric threshold used to create the rejection set.
}
\description{
This is a variable selection procedure that controls the false discovery rate, as proposed by Barber and Candes (2015)
}
