% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oatk.R
\name{bh}
\alias{bh}
\title{Benjamini-Hochberg procedure}
\usage{
bh(y, X, q = 0.1)
}
\arguments{
\item{y}{Response variable}

\item{X}{Design matrix}

\item{q}{The desired FDR level}
}
\value{
A list containing the vector rej, which are the indices of the columns of X corresponding to the non-null variables.
}
\description{
This function conducts the Benjamini-Hochberg procedure, which controls
the false discovery rate (FDR) in multiple hypothesis tests. The procedure is based on Benjamini and Hochberg (1995)
}
