% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{calculateFDP}
\alias{calculateFDP}
\title{calculateFDP}
\usage{
calculateFDP(beta_true, rej)
}
\arguments{
\item{beta_true}{The true regression coefficients.}

\item{rej}{The rejection set}
}
\value{
A numeric indicating the FDP.
}
\description{
Calculates the false discovery proportion.
}
