% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{calculateTDP}
\alias{calculateTDP}
\title{calculateTDP}
\usage{
calculateTDP(beta_true, rej)
}
\arguments{
\item{beta_true}{The true regression coefficients.}

\item{rej}{The rejection set}
}
\value{
A numeric indicating the TDP
}
\description{
Calculates the true discovery proportion.
}
