% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oatk.R
\name{calibrateOATK}
\alias{calibrateOATK}
\title{Conditionally calibrated one-at-a-time knockoffs (OATK)}
\usage{
calibrateOATK(y, X, lst.oatk, offset = 0, q = 0.1, M = 300L)
}
\arguments{
\item{y}{Response variable}

\item{X}{Design matrix}

\item{lst.oatk}{The output list from the basic oatk procedure ran using oatk()}

\item{offset}{Offset parameter when estimated the FDR. Setting c=1 results in a more conservative FDR result.}

\item{q}{The desired FDR level}

\item{M}{The number of replicates to run for each variable during the Monte Carlo simulations}
}
\value{
A list containing the vector rej, which are the indices of the columns of X corresponding to the non-null variables,
and the vector e_val which is the e-values for the variables.
}
\description{
This conducts conditional calibration on the results from the basic OATK procedure,
proposed by Guan, Ren, and Apley (2025). T. It achieves finite FDR control.
}
