% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{check.muts}
\alias{check.muts}
\title{check.muts}
\usage{
check.muts(muts.in)
}
\arguments{
\item{muts.in}{Input mutation labels}
}
\value{
Nothing if all mutations are correctly labeled. Otherwise, it throws an error.
}
\description{
Helper function to check that the mutations have been entered correctly.
Taken from https://hivdb.stanford.edu/download/GenoPhenoDatasets/DRMcv.R
}
