% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{convert.muts}
\alias{convert.muts}
\title{convert.muts}
\usage{
convert.muts(muts.in)
}
\arguments{
\item{muts.in}{Input mutation labels}
}
\value{
Corrected input mutation labels
}
\description{
Helper function to check that the mutations have been entered correctly
if the letters are entered as lower case, they are converted to upper case. Convert insertions or deletions to # or ~.
Taken from https://hivdb.stanford.edu/download/GenoPhenoDatasets/DRMcv.R
}
