% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{eBH}
\alias{eBH}
\title{eBH}
\usage{
eBH(e_values, alpha)
}
\arguments{
\item{e_values}{Numeric vector containing e-values}

\item{alpha}{The desired FDR level.}
}
\value{
A numeric vector corresponding to the rejection set.
}
\description{
This conducts the eBH procedure, proposed by Wang and Ramdas (2022),
to conduct variable selection with controlled false discovery rate (FDR) using e-values.
}
