% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{getDMCX}
\alias{getDMCX}
\title{getDMCX}
\usage{
getDMCX(n, p)
}
\arguments{
\item{n}{Number of rows of X}

\item{p}{Number of columns of X}
}
\value{
Design matrix X with n rows and p columns
}
\description{
generates a random design matrix X where each row is an independent sample
from a discrete-time Markov Chain model. For details, see Guan, Ren, and Apley (2025)
}
