% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{getGaussianX}
\alias{getGaussianX}
\title{getGaussianX}
\usage{
getGaussianX(n, p, rho = 0.4, mode = "power_decay")
}
\arguments{
\item{n}{Number of rows of X}

\item{p}{Number of columns of X}

\item{rho}{Scaling parameter of the covariance matrix}

\item{mode}{Covariance structure type. Takes either 'power_decay', 'const_pos', or 'const_neg'.
'power_decay' is the power decay covariance with decay rate rho.
'const_pos' is constant positive covariance with constant rho.
'const_neg' is constant negative covariance with constant rho in the precision matrix.}
}
\value{
Design matrix X with n rows and p columns
}
\description{
generates a random design matrix X from a multivariate normal with
zero mean and a covariance structure specified by the input arguments. See
Guan, Ren and Apley (2025) for further details.
}
