% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{getHIVData}
\alias{getHIVData}
\title{getHIVData}
\usage{
getHIVData(dataset = "PI", min.muts = 10)
}
\arguments{
\item{dataset}{string indicating which dataset to pull data from. Options are 'PI', 'NRTI', and 'NNRTI'.}

\item{min.muts}{numeric indicating how many times a mutation must occur in the dataset to be included in X}
}
\value{
Design matrix X
}
\description{
Creates a design matrix X from Stanford University's
HIV Drug Resistance Database. Adopted from https://hivdb.stanford.edu/download/GenoPhenoDatasets/DRMcv.R.
}
