% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{getHIVDataWithY}
\alias{getHIVDataWithY}
\title{getHIVDataWithY}
\usage{
getHIVDataWithY(dataset = "PI", min.muts = 10, drug = "FPV")
}
\arguments{
\item{dataset}{string indicating which dataset to pull data from. Options are 'PI', 'NRTI', and 'NNRTI'.}

\item{min.muts}{numeric indicating how many times a mutation must occur in the dataset to be included in X}

\item{drug}{string indicating which fold resistance data to collect for y.}
}
\value{
list with entry y (the response variable) and X (the design matrix)
}
\description{
Creates a matrix consisting of the response vector y corresponding
to the fold resistance data and a design matrix X from Stanford University's
HIV Drug Resistance Database. Adopted from https://hivdb.stanford.edu/download/GenoPhenoDatasets/DRMcv.R.
}
