% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{getToeplitzX}
\alias{getToeplitzX}
\title{getToeplitzX}
\usage{
getToeplitzX(n, p, rho = 0.5)
}
\arguments{
\item{n}{Number of rows of X}

\item{p}{Number of columns of X \strong{(must be divisible by 10)}.}

\item{rho}{Numeric. Value on the first off-diagonal inside each block; farther
off-diagonals decrease linearly to 0.}
}
\value{
Design matrix X with n rows and p columns
}
\description{
generates a random design matrix X from a multivariate normal with
zero mean and a covariance matrix that is Toeplitz (block diagonal). See
Guan, Ren and Apley (2025) for further details.
}
