% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{getY}
\alias{getY}
\title{getY}
\usage{
getY(X, beta_true, amp = 1)
}
\arguments{
\item{X}{Design matrix}

\item{beta_true}{The true coefficients of the model}

\item{amp}{Signal amplitude}
}
\value{
Numeric vector
}
\description{
Generates the response vector Y based on the linear model.
}
