% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oatk.R
\name{gm_low_d}
\alias{gm_low_d}
\title{Gaussian mirror}
\usage{
gm_low_d(y, X, q = 0.1, lam.lasso = NULL, offset = 0)
}
\arguments{
\item{y}{Response variable}

\item{X}{Design matrix}

\item{q}{The desired FDR level}

\item{lam.lasso}{The lasso parameter used to fit the regression model. If null, the function will find the optimal
value using cross-validation.}

\item{offset}{Offset parameter when estimated the FDR. Setting c=1 results in a more conservative FDR result.}
}
\value{
A list containing the vector rej, which are the indices of the columns of X corresponding to the non-null variables,
the vector W which are the test statistics for each variable, and the numeric threshold used to create the rejection set.
}
\description{
This is a variable selection procedure developed by Xing, Zhao, and Liu (2023)
that controls the false discovery rate. This implementation corresponds to Algorithm 1 in their paper.
However, instead of using OLS, we use Lasso regression, which boosts power. If you seek to use OLS, just enter lam.lasso=0.
}
