% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oatk.R
\name{knockDMC}
\alias{knockDMC}
\title{knockoffDMC}
\usage{
knockDMC(y, X, q = 0.1, offset = 0)
}
\arguments{
\item{y}{Response variable}

\item{X}{Design matrix}

\item{q}{The desired FDR level}

\item{offset}{Offset parameter when estimated the FDR. Setting c=1 results in a more conservative FDR result.}
}
\value{
A list containing the vector rej, which are the indices of the columns of X corresponding to the non-null variables,
the vector W which are the test statistics for each variable, and the numeric threshold used to create the rejection set.
}
\description{
This is a knockoff procedure specifically made for design matrix that is generated
from discrete Markov chains, developed by Sesia et al. (2019)
}
