% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oatk.R
\name{oatk}
\alias{oatk}
\title{One-at-a-time Knockoffs (OATK)}
\usage{
oatk(y, X, q = 0.1, lam.ridge = NULL, offset = 0, U = NULL, V = NULL, D = NULL)
}
\arguments{
\item{y}{Response variable}

\item{X}{Design matrix}

\item{q}{The desired FDR level}

\item{lam.ridge}{The ridge parameter used to fit the regression model. If null, the function will find the optimal
value using leave-one-out cross-validation.}

\item{offset}{Offset parameter when estimated the FDR. Setting c=1 results in a more conservative FDR result.}

\item{U}{the matrix U from the singular value decomposition of X=UDV^t. If any of U, D, or V is null, the function will compute them
using svd().}

\item{V}{the matrix V from the singular value decomposition of X=UDV^t}

\item{D}{the matrix D from the singular value decomposition of X=UDV^t}
}
\value{
A list containing the vector rej, which are the indices of the columns of X corresponding to the non-null variables,
the vector W which are the test statistics for each variable, the numeric threshold used to create the rejection set,
and numeric lam.ridge which was the used ridge parameter.
}
\description{
OATK is a variable selection procedure that controls the false discovery rate (FDR).
Proposed by Guan, Ren, and Apley (2025), it achieves higher power compared to other FDR-controlling
procedures such as Benjamini-Hochberg, knockoff filter, and Gaussian mirror. This implementation conducts
a fast version of OATK using ridge regression. If you desire to run OATK with OLS, set lam.ridge=0.
}
