% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oatk.R
\name{oatk_derandomized}
\alias{oatk_derandomized}
\title{Derandomized One-at-a-time Knockoffs (OATK)}
\usage{
oatk_derandomized(
  y,
  X,
  q = 0.1,
  lam.ridge = NULL,
  M = 31,
  eta = 0.5,
  offset = 0
)
}
\arguments{
\item{y}{Response variable}

\item{X}{Design matrix}

\item{q}{The desired FDR level}

\item{lam.ridge}{The ridge parameter used to fit the regression model. If null, the function will find the optimal
value using leave-one-out cross-validation.}

\item{M}{The number of replicates to run for each variable}

\item{eta}{The frequency threshold to include in the rejection set.}

\item{offset}{Offset parameter when estimated the FDR. Setting c=1 results in a more conservative FDR result.}
}
\value{
A list containing the vector rej, which are the indices of the columns of X corresponding to the non-null variables,
the matrix rej.mat which contains all the rejection results for each variable across the replicates.
}
\description{
This conducts the derandomized version of OATK, which is a variable selection procedure that controls the false discovery rate (FDR)
proposed by Guan, Ren, and Apley (2025). They showed derandomization reduces the variances of false discovery proportions and true discovery proportions
compared to the basic OATK.
}
