% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oatk.R
\name{oatk_multiple}
\alias{oatk_multiple}
\title{Multi-bit One-at-a-time Knockoffs (OATK)}
\usage{
oatk_multiple(
  y,
  X,
  q = 0.1,
  lam.ridge = NULL,
  M = 3L,
  gamma = max(1/(M + 1), q/(1 + q)),
  offset = 0
)
}
\arguments{
\item{y}{Response variable}

\item{X}{Design matrix}

\item{q}{The desired FDR level}

\item{lam.ridge}{The ridge parameter used to fit the regression model. If null, the function will find the optimal
value using leave-one-out cross-validation.}

\item{M}{The number of replicates to run for each variable. Higher M yields finer p-values.}

\item{gamma}{Rejection hyperparameter.}

\item{offset}{Offset parameter when estimated the FDR. Setting c=1 results in a more conservative FDR result.}
}
\value{
A list containing the vector rej, which are the indices of the columns of X corresponding to the non-null variables,
the vector p_val which is the p-values for each variable, and the vector stats which is the test statistics for each variable.
}
\description{
This conducts the multi-bit version of OATK, which is a variable selection procedure that controls the false discovery rate (FDR) and
produces p-values, proposed by Guan, Ren, and Apley (2025).
}
