% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{ridge.reg.Kfold}
\alias{ridge.reg.Kfold}
\title{Ridge regression}
\usage{
ridge.reg.Kfold(y, X, U_full, V_full, D_full, K = 10L, lambda_seq)
}
\arguments{
\item{y}{Response variable}

\item{X}{Design matrix}

\item{K}{The number of folds}

\item{lambda_seq}{Ridge parameter. If lam is a scalar, ridge will be conducted with that lam.}
}
\value{
A list containing beta_hat (the regression coefficients) and lam.ridge (the corresponding ridge parameter)
}
\description{
Function for conducting ridge and for finding lambda using K-fold cros-validation if
lambda is a vector.
}
