% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{ridge.reg}
\alias{ridge.reg}
\title{Ridge regression}
\usage{
ridge.reg(Y, U, D, V, lam)
}
\arguments{
\item{Y}{Response variable}

\item{U}{U in the SVD X = UDV^t,}

\item{D}{D in the SVD X = UDV^t,}

\item{V}{V in the SVD X = UDV^t,}

\item{lam}{Ridge parameter. If lam is a scalar, ridge will be conducted with that lam.
(X nearly singular & training fit nearly perfect),}
}
\value{
A list containing beta_hat (the regression coefficients) and lam.ridge (the corresponding ridge parameter)
}
\description{
Function for conducting ridge and for finding lambda using fast leave-one-out cross-validation (LOOCV) if
lambda is a vector.
}
