% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{sampleNonnull}
\alias{sampleNonnull}
\title{sampleNonnull}
\usage{
sampleNonnull(p, k)
}
\arguments{
\item{p}{Numeric indicating length of vector}

\item{k}{Numeric indicating number of nonzero entries}
}
\value{
A vector
}
\description{
Creates a random coefficient vector with k nonzero entries, whose signs
are sampled uniformly at random.
}
